move clibot parsing args sys.argv, etc to constants
