+++ data/corpora/data_science/glossary.md
@@ -1,0 +1,9 @@
+precision:
+recall:
+accuracy:
+f1 score:
+metric:
+measure:
+standard deviation:
+variance:
+root mean square error (rmse): often pronounced "rem-zee"
