+++ nlpia-bot.ini
@@ -5,6 +5,7 @@ num_top_replies: 10
 bots: parul,eliza,pattern,glossary  # bots: glossary,pattern,parul,search_fuzzy,eliza,time
 # loglevel: 10=debug 20=info 30=warn 40=error 50=fatal
 loglevel: 40
+lang: en_core_web_lg
 # quality_weights: {'spell': .5, 'semantic': .5}
 # loglevel: WARNING
 # self_score: .5
