+++ setup.cfg
@@ -67,6 +67,7 @@ install_requires =
     tensorboard
     transformers
     wikipedia-api
+    environment
     # wrapt<1.11.0,>=1.10.0
 
 # The usage of test_requires is discouraged, see `Dependency Management` docs
